# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/invoices/{id}/pdf": {
      "get": {
        "operationId": "BillingController_pdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["billing"]
      }
    },
    "/api/invoices/{id}/payments": {
      "post": {
        "operationId": "BillingController_recordPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordPaymentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["billing"]
      }
    },
...
    "/api/me/invoices/{id}/pdf": {
      "get": {
        "operationId": "MeController_invoicePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["me"]
      }
    },
...
    "/api/dental-lab/invoices": {
      "get": {
        "operationId": "DentalLabInvoicesController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DentalLabInvoiceStatus"
            }
          },
          {
            "name": "clinicTenantId",
            "required": false,
            "in": "query",
            "description": "Filter to a specific clinic (lab-side only).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 500,
              "default": 200,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      },
      "post": {
        "operationId": "DentalLabInvoicesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLabInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}": {
      "get": {
        "operationId": "DentalLabInvoicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      },
      "patch": {
        "operationId": "DentalLabInvoicesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/generate-from-cases": {
      "post": {
        "operationId": "DentalLabInvoicesController_generate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateFromCasesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/sweep": {
      "post": {
        "operationId": "DentalLabInvoicesController_sweep",
        "parameters": [
          {
            "name": "period",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/items": {
      "post": {
        "operationId": "DentalLabInvoicesController_addItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddLabInvoiceItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/items/{itemId}": {
      "patch": {
        "operationId": "DentalLabInvoicesController_updateItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      },
      "delete": {
        "operationId": "DentalLabInvoicesController_deleteItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/issue": {
      "post": {
        "operationId": "DentalLabInvoicesController_issue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/payments": {
      "post": {
        "operationId": "DentalLabInvoicesController_recordPayment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordLabInvoicePaymentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/void": {
      "post": {
        "operationId": "DentalLabInvoicesController_voidInvoice",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/pdf": {
      "post": {
        "operationId": "DentalLabInvoicesController_generatePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/payment-links": {
      "post": {
        "operationId": "DentalLabInvoicesController_createPaymentLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentLinkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
    "/api/dental-lab/invoices/{id}/payment-links/{linkId}": {
      "delete": {
        "operationId": "DentalLabInvoicesController_cancelPaymentLink",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["lab-invoices"]
      }
    },
...
    "/api/clinic/lab-invoices/{id}": {
      "get": {
        "operationId": "ClinicLabInvoicesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-invoices"]
      }
    },
    "/api/clinic/lab-invoices/{id}/pdf": {
      "get": {
        "operationId": "ClinicLabInvoicesController_getPdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": ["clinic-lab-invoices"]
      }
    },
...
  },
  "info": {
    "title": "ClinIQ API",
    "description": "Multi-tenant clinic management API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4005"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RecordPaymentDto": {
        "type": "object",
        "properties": {
          "amountCentavos": {
            "type": "number",
            "description": "Amount in centavos"
          },
          "method": {
            "type": "string",
            "enum": [
              "CASH",
              "GCASH",
              "MAYA",
              "BANK_TRANSFER",
              "CARD",
              "HMO",
              "INSURANCE",
              "OTHER"
            ]
          },
          "reference": {
            "type": "string",
            "description": "Gateway / OR reference"
          }
        },
        "required": ["amountCentavos", "method"]
      },
...
      "DentalLabInvoiceStatus": {
        "type": "string",
        "enum": ["DRAFT", "ISSUED", "PAID", "OVERDUE", "VOID"]
      },
      "DentalLabInvoiceItemInputDto": {
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "description": "Optional DentalLabCase id this line bills for."
          },
          "description": {
            "type": "string",
            "example": "Crown — case #142"
          },
          "qty": {
            "type": "number",
            "example": 1,
            "default": 1
          },
          "unitPriceCents": {
            "type": "number",
            "description": "Unit price in centavos.",
            "example": 250000
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          }
        },
        "required": ["description", "unitPriceCents"]
      },
      "CreateLabInvoiceDto": {
        "type": "object",
        "properties": {
          "clinicTenantId": {
            "type": "string",
            "description": "Clinic tenant id this invoice is billed to."
          },
          "currency": {
            "type": "string",
            "default": "PHP"
          },
          "dueAt": {
            "type": "string",
            "description": "ISO 8601 due date."
          },
          "notes": {
            "type": "string"
          },
          "taxCents": {
            "type": "number",
            "description": "Tax in centavos (line-total agnostic).",
            "default": 0
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DentalLabInvoiceItemInputDto"
            }
          }
        },
        "required": ["clinicTenantId"]
      },
      "GenerateFromCasesDto": {
        "type": "object",
        "properties": {
          "clinicTenantId": {
            "type": "string",
            "description": "Clinic tenant id. All cases must belong to this clinic."
          },
          "caseIds": {
            "description": "DentalLabCase ids to bill.",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "currency": {
            "type": "string",
            "default": "PHP"
          },
          "dueAt": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "taxCents": {
            "type": "number",
            "default": 0
          }
        },
        "required": ["clinicTenantId", "caseIds"]
      },
      "UpdateInvoiceDto": {
        "type": "object",
        "properties": {
          "dueAt": {
            "type": "object"
          },
          "notes": {
            "type": "object"
          },
          "taxCents": {
            "type": "number",
            "default": 0
          }
        }
      },
      "AddLabInvoiceItemDto": {
        "type": "object",
        "properties": {
          "caseId": {
            "type": "string",
            "description": "Optional DentalLabCase id this line bills for."
          },
          "description": {
            "type": "string",
            "example": "Crown — case #142"
          },
          "qty": {
            "type": "number",
            "example": 1,
            "default": 1
          },
          "unitPriceCents": {
            "type": "number",
            "description": "Unit price in centavos.",
            "example": 250000
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          }
        },
        "required": ["description", "unitPriceCents"]
      },
      "UpdateInvoiceItemDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "qty": {
            "type": "number"
          },
          "unitPriceCents": {
            "type": "number"
          },
          "sortOrder": {
            "type": "number"
          }
        }
      },
      "RecordLabInvoicePaymentDto": {
        "type": "object",
        "properties": {
          "amountCents": {
            "type": "number",
            "description": "Amount paid in centavos. Adds to paidCents."
          },
          "paidAt": {
            "type": "string",
            "description": "ISO 8601 timestamp of the payment."
          },
          "reference": {
            "type": "string",
            "description": "Free-text reference (bank ref, link id, etc.)."
          }
        },
        "required": ["amountCents"]
      },
      "DentalLabPaymentLinkProvider": {
        "type": "string",
        "enum": ["PAYMONGO", "GCASH", "MAYA", "STRIPE", "MANUAL"]
      },
      "CreatePaymentLinkDto": {
        "type": "object",
        "properties": {
          "provider": {
            "default": "MANUAL",
            "allOf": [
              {
                "$ref": "#/components/schemas/DentalLabPaymentLinkProvider"
              }
            ]
          },
          "amountCents": {
            "type": "number",
            "description": "Amount in centavos. Defaults to invoice outstanding balance."
          },
          "externalId": {
            "type": "string",
            "description": "External provider id (PayMongo link id, etc.)."
          },
          "url": {
            "type": "string",
            "description": "Public-facing URL the payer follows."
          },
          "expiresAt": {
            "type": "string",
            "description": "ISO 8601 expiry."
          }
        }
      },
...
    }
  }
}